pyproject.toml
[project]
name = "super-octo-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "realmordor", email = "realmordor@github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "zeep>=4.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
